data/code/requirements.txt
rasterio
albumentations
segmentation_models_pytorch
tqdm
torch
numpy
Pillow
torchvision